# docs/architecture/api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "A股/港股/美股自选股智能分析系统 API\n\n## 功能模块\n- 股票分析：触发 AI 智能分析\n- 历史记录：查询历史分析报告\n- 股票数据：获取行情数据\n\n## 认证方式\n支持可选管理员认证：ADMIN_AUTH_ENABLED=true 时，除登录、状态、健康检查和 OpenAPI 文档外，/api/v1/* 需要有效管理员会话 Cookie；关闭时不强制认证。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
...
    "/api/v1/stocks/{stock_code}/profile": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "summary": "获取个股研究聚合档案",
        "description": "按独立质量状态聚合行情、历史、研究产物、资讯、持仓关系和监控规则。",
        "operationId": "get_stock_profile_api_v1_stocks__stock_code__profile_get",
        "parameters": [
          {
            "name": "stock_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stock Code"
            }
          },
          {
            "name": "history_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "日线历史天数",
              "default": 60,
              "title": "History Days"
            },
            "description": "日线历史天数"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "股票代码无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{stock_code}/kline": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "summary": "获取 K 线图数据（含均线与缠论笔/线段）",
        "description": "返回指定周期的 K 线、均线和缠论笔/线段端点。daily/weekly 复用日线多数据源回退（周线由日线聚合）；60m/30m 仅支持 A 股，来自 AkShare 新浪分钟线（前复权）。",
        "operationId": "get_stock_kline_api_v1_stocks__stock_code__kline_get",
        "parameters": [
          {
            "name": "stock_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stock Code"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(daily|weekly|60m|30m)$",
              "description": "K 线周期",
              "default": "daily",
              "title": "Period"
            },
            "description": "K 线周期"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1500,
              "minimum": 30,
              "description": "日线/周线回看的交易日数量（分钟线忽略）",
              "default": 365,
              "title": "Days"
            },
            "description": "日线/周线回看的交易日数量（分钟线忽略）"
          }
        ],
        "responses": {
          "200": {
            "description": "K 线图数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockKlineResponse"
                }
              }
            }
          },
          "422": {
            "description": "不支持的代码或周期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "数据源获取失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "错误类型",
            "example": "validation_error"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "错误详情",
            "example": "请求参数错误"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "附加错误信息"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "错误响应",
        "example": {
          "error": "not_found",
          "message": "资源不存在"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StockProfileResearchBlock": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fresh",
              "partial",
              "unavailable"
            ],
            "title": "Status"
          },
          "data": {
            "$ref": "#/components/schemas/StockProfileResearchData"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StockProfileResearchBlock"
      },
...
      "StockProfileHistoryBlock": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fresh",
              "partial",
              "unavailable"
            ],
            "title": "Status"
          },
          "period": {
            "type": "string",
            "const": "daily",
            "title": "Period",
            "default": "daily"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/KLineData"
            },
            "type": "array",
            "title": "Data"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StockProfileHistoryBlock"
      },
...
      "StockProfilePortfolioBlock": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fresh",
              "partial",
              "unavailable"
            ],
            "title": "Status"
          },
          "data": {
            "$ref": "#/components/schemas/StockProfilePortfolioRelation"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StockProfilePortfolioBlock"
      },
...
      "StockProfileMonitorBlock": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fresh",
              "partial",
              "unavailable"
            ],
            "title": "Status"
          },
          "data": {
            "$ref": "#/components/schemas/StockProfileMonitorData"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StockProfileMonitorBlock"
      },
...
      "StockProfileIntelligenceBlock": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fresh",
              "partial",
              "unavailable"
            ],
            "title": "Status"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/IntelligenceItem"
            },
            "type": "array",
            "title": "Items"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StockProfileIntelligenceBlock"
      },
...
      "StockProfileResponse": {
        "properties": {
          "requested_code": {
            "type": "string",
            "title": "Requested Code"
          },
          "canonical_code": {
            "type": "string",
            "title": "Canonical Code"
          },
          "market": {
            "type": "string",
            "enum": [
              "cn",
              "hk",
              "us",
              "jp",
              "kr",
              "tw"
            ],
            "title": "Market"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "quote": {
            "$ref": "#/components/schemas/StockProfileQuoteBlock"
          },
          "history": {
            "$ref": "#/components/schemas/StockProfileHistoryBlock"
          },
          "research": {
            "$ref": "#/components/schemas/StockProfileResearchBlock"
          },
          "intelligence": {
            "$ref": "#/components/schemas/StockProfileIntelligenceBlock"
          },
          "portfolio": {
            "$ref": "#/components/schemas/StockProfilePortfolioBlock"
          },
          "monitors": {
            "$ref": "#/components/schemas/StockProfileMonitorBlock"
          },
          "evidence_quality": {
            "$ref": "#/components/schemas/StockProfileEvidenceQuality"
          }
        },
        "type": "object",
        "required": [
          "requested_code",
          "canonical_code",
          "market",
          "as_of",
          "quote",
          "history",
          "research",
          "intelligence",
          "portfolio",
          "monitors",
          "evidence_quality"
        ],
        "title": "StockProfileResponse"
      },
      "StockProfileQuoteBlock": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fresh",
              "partial",
              "unavailable"
            ],
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StockQuote"
              },
              {
                "type": "null"
              }
            ]
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StockProfileQuoteBlock"
      },
...
      "StockKlineResponse": {
        "properties": {
          "stock_code": {
            "type": "string",
            "title": "Stock Code",
            "description": "股票代码"
          },
          "stock_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Name",
            "description": "股票名称"
          },
          "period": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "60m",
              "30m"
            ],
            "title": "Period",
            "description": "K 线周期"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "数据源"
          },
          "bars": {
            "items": {
              "$ref": "#/components/schemas/ChartBar"
            },
            "type": "array",
            "title": "Bars",
            "description": "K 线"
          },
          "moving_averages": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "type": "object",
            "title": "Moving Averages",
            "description": "均线，键为周期，值与 bars 一一对应"
          },
          "chan": {
            "$ref": "#/components/schemas/ChanStructure",
            "description": "缠论结构"
          }
        },
        "type": "object",
        "required": [
          "stock_code",
          "period"
        ],
        "title": "StockKlineResponse",
        "description": "K 线图数据：多周期 K 线 + 均线 + 缠论笔/线段"
      },
...
    },
    "securitySchemes": {
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dsa_session"
      }
    }
  }
}